requirements.txt
assemblyai
deep-translator
google-generativeai
elevenlabs
python-dotenv
sounddevice
numpy
streamlit
langdetect
flask
flask-cors